+++ tidy3d/components/geometry/utils_2d.py
@@ -172,7 +172,6 @@ def subdivide(
         # If the 2D structure overlaps completely with all previously tested structures above then there is no more work to do
         if not geom_shapely:
             break
-
         intersection_res = shapely.intersection(geom_shapely, mp_structure[0])
         intersection_mp = to_multipolygon(intersection_res)
         difference_res = shapely.difference(geom_shapely, mp_structure[0])
@@ -194,7 +193,6 @@ def subdivide(
                 break
             if not mp_structure_below[0]:
                 continue
-
             intersection_res = shapely.intersection(above_intersection, mp_structure_below[0])
             intersection_mp = to_multipolygon(intersection_res)
             above_difference = to_multipolygon(